+++ engine-io-snap/Changelog.md
@@ -2,6 +2,8 @@
 
 * Updated for the new `ServerAPI` in `engine-io-1.1`.
 
+* Fixed incomplete pattern matches when parsing request method.
+
 ## 1.0.0
 
 * Initial release.