+++ engine-io-snap/Changelog.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+* Updated for `ServerAPI` in `engine-io-1.2`.
+
+* ParseError is now caught correctly, and no longer results in a response 500 to the client.
+
 ## 1.0.1
 
 * Updated for the new `ServerAPI` in `engine-io-1.1`.